api/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
python-multipart==0.0.6
transformers==4.30.2
torch==2.0.1
docx2txt==0.8
PyMuPDF==1.22.1
markdown==3.4.3
python-pptx==0.6.21
fpdf==1.7.2